Refactored AstCreator and friends
This is a pure structural refactoring. No changes in semantics.
Simply speaking, the >1000 loc AstCreator was split into individual creators
responsible for various C/C++ AST node types (expressions, statements, types, etc.).
