+++ c2cpg/src/main/scala/io/shiftleft/c2cpg/passes/AstCreationPass.scala
@@ -1,6 +1,8 @@
 package io.shiftleft.c2cpg.passes
 
 import io.shiftleft.c2cpg.C2Cpg
+import io.shiftleft.c2cpg.astcreation.AstCreator
+import io.shiftleft.c2cpg.datastructures.Global
 import io.shiftleft.c2cpg.parser.CdtParser.ParseResult
 import io.shiftleft.c2cpg.parser.{CdtParser, ParseConfig}
 import io.shiftleft.codepropertygraph.Cpg
